% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histdata.R
\docType{data}
\name{histdata}
\alias{histdata}
\title{Kenya Exchange Rate Historical Data.}
\format{
A data frame with 9373 observations of 5 variables.
\describe{
\item{Date}{Date corresponding to day, month and year}
\item{Currency}{Specifying the currency in consideration}
\item{Mean}{The average of the Buy and Sell columns}
\item{Sell}{The price of selling US Dollar to the bank}
\item{Buy}{The price of buying US Dollar from the bank}
}
}
\source{
https://www.centralbank.go.ke/rates/forex-exchange-rates/
}
\usage{
histdata
}
\description{
Kenya Foreign Exchange data from 11 Oct 2016 to 17 February 2022 for 7 currencies (US DOLLAR, EURO STG POUND, BIF, TSHS, USHS, RWF).
}
\keyword{datasets}
